+++ priv/repo/migrations/20190816113332_add_report_count_to_projects.exs
@@ -1,0 +1,9 @@
+defmodule Birdiy.Repo.Migrations.AddReportCountToProjects do
+  use Ecto.Migration
+
+  def change do
+    alter table(:projects) do
+      add :report_count, :integer, null: false, default: 0
+    end
+  end
+end