+++ priv/repo/migrations/20190816113418_add_report_count_to_posts.exs
@@ -1,0 +1,9 @@
+defmodule Birdiy.Repo.Migrations.AddReportCountToPosts do
+  use Ecto.Migration
+
+  def change do
+    alter table(:posts) do
+      add :report_count, :integer, null: false, default: 0
+    end
+  end
+end