+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Pages of categories and topics.
 - Sitemap.
+- Post reporting.
+- Project reporting.
 
 ## [0.0.7] - 2019-08-13
 ### Added